Separate plotting FigureWindow into a separate module
Refs #0
